Add Gemfile for dependencies and update how get setup
Whilst I was making a single line contribution to this, I found I was missing
gems and there was not an automated way of getting setup. To rectify this, I
have added a Gemfile and notes in the contributing guidelines for users to run
`bundle install` instead of needing to go through the require statements and
install them manually.