Changed title back to nickname
Thanks for the effort, @duckspeaker - but the official name of my blog is "Leo" and not "Leonard Lamprecht".

I know this might be a bit confusing (particularly because I didn't change the logo text yet), but there are some good reasons I didn't name it after my full name, primarily because it's just too hard to pronounce.

I already changed my domain and the <title> tag to the new name and will also make sure the logo fits it. Sorry for the confusion! :)